Fix bug in entry size calculation

+++ recall-store/src/main/java/com/aitusoftware/recall/map/ByteSequenceMap.java
@@ -82,7 +82,7 @@ public final class ByteSequenceMap implements SequenceMap<ByteBuffer>
     {
         totalEntryCount = BitUtil.findNextPositivePowerOfTwo(initialSize);
         entryMask = totalEntryCount - 1;
-        entrySizeInBytes = (maxKeyLength + Integer.BYTES + Long.BYTES);
+        entrySizeInBytes = (maxKeyLength + Long.BYTES + Long.BYTES);
         final long bufferSize = entrySizeInBytes * (long)totalEntryCount;
         if (bufferSize > ((long)Integer.MAX_VALUE) || totalEntryCount < 0)
         {
